identify frequent routes
I was able to identify the frequency that Bay Wheels trips go down unique street segments (in the East Bay region only so far) using the gpd.tools.sjoin() method. This intersected the segments within the routes so I could append the counts column to the segments gdf. Afterwords the dataset is grouped to get total trips per segment and mapped.